+++ CHANGELOG.md
@@ -15,6 +15,10 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 
 ## [Unreleased](https://github.com/hynek/svcs/compare/23.13.0...HEAD)
 
+### Added
+
+- aiohttp: missing `aget_abstract()` function.
+
 
 ## [23.13.0](https://github.com/hynek/svcs/compare/23.12.0...23.13.0) - 2023-08-11